# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: FitLog API
  description: '

    FitLog は筋トレ・運動記録サービスの REST API です。


    ## 認証

    保護エンドポイントは Bearer トークン（JWT）を要求します。

    `/auth/login` で取得したトークンを `Authorization: Bearer <token>` ヘッダに付与してください。


    ## 共通エラーレスポンス

    | ステータス | 意味 |

    |---|---|

    | 401 | 未認証（トークン不正・期限切れ） |

    | 403 | 権限なし（他ユーザーのリソースへのアクセス等） |

    | 404 | リソースが見つからない |

    | 409 | 重複登録（メール・ユーザー名等） |

    | 422 | バリデーションエラー |

    '
  version: 0.1.0
paths:
...
  /workouts/{workout_id}/cheers:
    post:
      tags:
      - cheers
      summary: ナイストレを付与
      description: 指定ワークアウットにナイストレ（いいね）を付与します。自分のワークアウットには付与できません（F-04）。
      operationId: add_cheer_workouts__workout_id__cheers_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workout_id
        in: path
        required: true
        schema:
          type: integer
          title: Workout Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheerStateOut'
        '401':
          description: 未認証
        '403':
          description: 自分のワークアウットへのナイストレは不可
        '404':
          description: ワークアウットが見つからない
        '409':
          description: 既にナイストレ済み
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - cheers
      summary: ナイストレを解除
      description: 指定ワークアウットに付与済みのナイストレを解除します（F-04）。
      operationId: remove_cheer_workouts__workout_id__cheers_delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workout_id
        in: path
        required: true
        schema:
          type: integer
          title: Workout Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 未認証
        '403':
          description: 自分のワークアウットへの操作は不可
        '404':
          description: ワークアウットが見つからない、またはナイストレしていない
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workouts/{workout_id}/advices:
    post:
      tags:
      - advices
      summary: アドバイスを投稿
      description: 指定ワークアウットにアドバイスコメントを投稿します。自分のワークアウットへは投稿できません（F-05）。
      operationId: create_advice_workouts__workout_id__advices_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workout_id
        in: path
        required: true
        schema:
          type: integer
          title: Workout Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdviceCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdviceOut'
        '401':
          description: 未認証
        '403':
          description: 自分のワークアウットへのアドバイスは不可
        '404':
          description: ワークアウットが見つからない
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - advices
      summary: ワークアウットのアドバイス一覧取得
      description: 指定ワークアウットに投稿されたアドバイスを投稿日時昇順で返します（F-05）。
      operationId: list_advices_workouts__workout_id__advices_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workout_id
        in: path
        required: true
        schema:
          type: integer
          title: Workout Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdviceOut'
                title: Response List Advices Workouts  Workout Id  Advices Get
        '401':
          description: 未認証
        '404':
          description: ワークアウットが見つからない
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /advices/{advice_id}:
    delete:
      tags:
      - advices
      summary: アドバイスを削除（投稿者のみ可）
      description: 自分が投稿したアドバイスを削除します。他ユーザーのアドバイスは削除できません（F-05）。
      operationId: delete_advice_advices__advice_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: advice_id
        in: path
        required: true
        schema:
          type: integer
          title: Advice Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 未認証
        '403':
          description: 他ユーザーのアドバイスへの操作は不可
        '404':
          description: アドバイスが見つからない
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /goals:
    get:
      tags:
      - goals
      summary: 目標一覧取得（週次・月次）
      description: ログインユーザーが設定した週次・月次の目標一覧を返します（F-07）。
      operationId: list_goals_goals_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GoalOut'
                type: array
                title: Response List Goals Goals Get
        '401':
          description: 未認証
      security:
      - OAuth2PasswordBearer: []
    put:
      tags:
      - goals
      summary: 目標を設定（upsert）
      description: 期間種別・指標・目標値で目標を設定します。既存の目標がある場合は上書き（upsert）します（F-07）。
      operationId: set_goal_goals_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalOut'
        '401':
          description: 未認証
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /goals/{period_type}:
    delete:
      tags:
      - goals
      summary: 目標を削除
      description: 指定した期間種別（weekly / monthly）の目標を削除します（F-07）。
      operationId: delete_goal_goals__period_type__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: period_type
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/PeriodTypePath'
      responses:
        '204':
          description: Successful Response
        '401':
          description: 未認証
        '404':
          description: 指定した期間の目標が設定されていない
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /goals/{period_type}/trend:
    get:
      tags:
      - goals
      summary: 目標達成率の推移データ取得
      description: 指定した期間種別の過去N期間分（デフォルト8）の達成率推移を返します。グラフ表示に使用します（F-07）。
      operationId: goal_trend_goals__period_type__trend_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: period_type
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/PeriodTypePath'
      - name: count
        in: query
        required: false
        schema:
          type: integer
          maximum: 24
          minimum: 1
          description: 取得する期間数（1〜24、デフォルト8）
          default: 8
          title: Count
        description: 取得する期間数（1〜24、デフォルト8）
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AchievementPointOut'
                title: Response Goal Trend Goals  Period Type  Trend Get
        '401':
          description: 未認証
        '404':
          description: 指定した期間の目標が設定されていない
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /dashboard/personal-records:
    get:
      tags:
      - dashboard
      summary: 種目別の自己ベスト一覧取得
      description: ログインユーザーの種目別自己ベスト（最大重量・ベストボリューム・推定1RM）を返します（F-09）。
      operationId: personal_records_dashboard_personal_records_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PersonalRecordOut'
                type: array
                title: Response Personal Records Dashboard Personal Records Get
        '401':
          description: 未認証
      security:
      - OAuth2PasswordBearer: []
  /dashboard/achievements:
    get:
      tags:
      - dashboard
      summary: 現在期間の目標達成率取得
      description: 今週・今月の目標達成率（実績値・達成率・達成済みフラグ）を返します（F-07）。
      operationId: achievements_dashboard_achievements_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AchievementOut'
                type: array
                title: Response Achievements Dashboard Achievements Get
        '401':
          description: 未認証
      security:
      - OAuth2PasswordBearer: []
  /dashboard/streak:
    get:
      tags:
      - dashboard
      summary: 連続記録日数（現在・最長）取得
      description: ログインユーザーの現在の連続記録日数と過去最長の連続記録日数を返します（F-08）。
      operationId: streak_dashboard_streak_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreakOut'
        '401':
          description: 未認証
      security:
      - OAuth2PasswordBearer: []
  /dashboard/heatmap:
    get:
      tags:
      - dashboard
      summary: 日別ボリュームヒートマップデータ取得
      description: 直近Nか月（デフォルト5）の日別ボリュームを返します。記録なし日も volume=0 で含まれます（F-08）。
      operationId: heatmap_dashboard_heatmap_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: months
        in: query
        required: false
        schema:
          type: integer
          maximum: 12
          minimum: 1
          description: 取得する月数（1〜12、デフォルト5）
          default: 5
          title: Months
        description: 取得する月数（1〜12、デフォルト5）
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HeatmapCell'
                title: Response Heatmap Dashboard Heatmap Get
        '401':
          description: 未認証
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AchievementOut:
      properties:
        period_type:
          type: string
          enum:
          - weekly
          - monthly
          title: Period Type
          description: 期間種別
        metric:
          type: string
          enum:
          - sessions
          - volume
          title: Metric
          description: 計測指標
        target_value:
          type: string
          title: Target Value
          description: 目標値
        actual:
          type: string
          title: Actual
          description: 現在の実績値
        rate:
          type: integer
          title: Rate
          description: 達成率（0〜100、超過は100に頭打ち）
        achieved:
          type: boolean
          title: Achieved
          description: 達成済みか（超過も達成扱い）
        period_start:
          type: string
          format: date
          title: Period Start
          description: 期間開始日（含む）
        period_end:
          type: string
          format: date
          title: Period End
          description: 期間終了日（排他、次期間の開始日）
      type: object
      required:
      - period_type
      - metric
      - target_value
      - actual
      - rate
      - achieved
      - period_start
      - period_end
      title: AchievementOut
      description: 現在期間（今週/今月）の達成状況。
    AchievementPointOut:
      properties:
        period_start:
          type: string
          format: date
          title: Period Start
          description: 期間開始日
        actual:
          type: string
          title: Actual
          description: 実績値
        target_value:
          type: string
          title: Target Value
          description: 目標値
        rate:
          type: integer
          title: Rate
          description: 達成率（0〜100）
        achieved:
          type: boolean
          title: Achieved
          description: 達成済みか
      type: object
      required:
      - period_start
      - actual
      - target_value
      - rate
      - achieved
      title: AchievementPointOut
      description: 達成率推移グラフの1点（期間バケット）。
    AdviceCreate:
      properties:
        content:
          type: string
          maxLength: 140
          minLength: 1
          title: Content
          description: アドバイス本文（1〜140文字）
          examples:
          - フォームが綺麗でした！
      type: object
      required:
      - content
      title: AdviceCreate
      description: 投稿入力。本文は必須・最大140文字（DBは LENGTH>=1 のみ、上限はここで担保）。
    AdviceOut:
      properties:
        id:
          type: integer
          title: Id
          description: アドバイスID
        workout_id:
          type: integer
          title: Workout Id
          description: 対象ワークアウトID
        user_id:
          type: integer
          title: User Id
          description: 投稿者ユーザーID
        username:
          type: string
          title: Username
          description: 投稿者ユーザー名
        content:
          type: string
          title: Content
          description: アドバイス本文
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 投稿日時（UTC）
      type: object
      required:
      - id
      - workout_id
      - user_id
      - username
      - content
      - created_at
      title: AdviceOut
      description: '一覧/詳細用。投稿者名・本文・日時を含む（要件: 昇順表示）。'
      example:
        content: フォームが綺麗でした！
        created_at: '2026-05-18T12:00:00'
        id: 1
        user_id: 2
        username: bob
        workout_id: 10
...
    CheerStateOut:
      properties:
        cheers_count:
          type: integer
          title: Cheers Count
          description: 現在のナイストレ総数
        cheered_by_me:
          type: boolean
          title: Cheered By Me
          description: ログインユーザーが付与済みか
      type: object
      required:
      - cheers_count
      - cheered_by_me
      title: CheerStateOut
      description: 付与/解除後の最新状態（件数 + 自分が付与済みか）。
...
    GoalIn:
      properties:
        period_type:
          type: string
          enum:
          - weekly
          - monthly
          title: Period Type
          description: 期間種別（weekly / monthly）
          examples:
          - weekly
        metric:
          type: string
          enum:
          - sessions
          - volume
          title: Metric
          description: 計測指標（sessions=回数 / volume=総ボリュームkg）
          examples:
          - sessions
        target_value:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
          title: Target Value
          description: 目標値（0より大きい数値）
          examples:
          - 3
      type: object
      required:
      - period_type
      - metric
      - target_value
      title: GoalIn
      description: 目標設定入力。period_type/metric は固定語彙、target_value は正の数。
    GoalOut:
      properties:
        id:
          type: integer
          title: Id
          description: 目標ID
        period_type:
          type: string
          enum:
          - weekly
          - monthly
          title: Period Type
          description: 期間種別
        metric:
          type: string
          enum:
          - sessions
          - volume
          title: Metric
          description: 計測指標
        target_value:
          type: string
          title: Target Value
          description: 目標値
        start_on:
          type: string
          format: date
          title: Start On
          description: 目標開始日
      type: object
      required:
      - id
      - period_type
      - metric
      - target_value
      - start_on
      title: GoalOut
      example:
        id: 1
        metric: sessions
        period_type: weekly
        start_on: '2026-05-18'
        target_value: '3'
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HeatmapCell:
      properties:
        date:
          type: string
          format: date
          title: Date
          description: 日付（YYYY-MM-DD）
        volume:
          type: string
          title: Volume
          description: その日の総ボリューム（kg × reps の合計、記録なし=0）
      type: object
      required:
      - date
      - volume
      title: HeatmapCell
      description: F-08 ヒートマップの1日。記録なし日も volume=0 で出力。
...
    PeriodTypePath:
      type: string
      enum:
      - weekly
      - monthly
      title: PeriodTypePath
    PersonalRecordOut:
      properties:
        exercise_id:
          type: integer
          title: Exercise Id
          description: 種目ID
        exercise_name:
          type: string
          title: Exercise Name
          description: 種目名
        max_weight_kg:
          anyOf:
          - type: string
          - type: 'null'
          title: Max Weight Kg
          description: 最大重量（kg）
        max_weight_reps:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Weight Reps
          description: 最大重量時のレップ数
        max_weight_on:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Max Weight On
          description: 最大重量を達成した日
        best_volume:
          anyOf:
          - type: string
          - type: 'null'
          title: Best Volume
          description: 1日の最大総ボリューム（kg × reps）
        best_volume_on:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Best Volume On
          description: ベストボリュームを達成した日
        best_est_1rm:
          anyOf:
          - type: string
          - type: 'null'
          title: Best Est 1Rm
          description: 推定1RM最大値（kg）
        best_1rm_weight_kg:
          anyOf:
          - type: string
          - type: 'null'
          title: Best 1Rm Weight Kg
          description: 推定1RM算出時の重量（kg）
        best_1rm_reps:
          anyOf:
          - type: integer
          - type: 'null'
          title: Best 1Rm Reps
          description: 推定1RM算出時のレップ数
        best_1rm_on:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Best 1Rm On
          description: 推定1RMベストを達成した日
      type: object
      required:
      - exercise_id
      - exercise_name
      - max_weight_kg
      - max_weight_reps
      - max_weight_on
      - best_volume
      - best_volume_on
      - best_est_1rm
      - best_1rm_weight_kg
      - best_1rm_reps
      - best_1rm_on
      title: PersonalRecordOut
      description: 種目別の自己ベスト（F-09 ダッシュボード併記表示）。
...
    StreakOut:
      properties:
        current:
          type: integer
          title: Current
          description: 現在の連続記録日数
        longest:
          type: integer
          title: Longest
          description: 過去最長の連続記録日数
      type: object
      required:
      - current
      - longest
      title: StreakOut
      description: F-08 ストリーク（現在の連続記録日数・最長）。
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login
...